♻️ extract session store from session management
